Integer and float members
C program to design a structure named XYZ with one integer and one float member.
Then declare 3 variables V1, V2 and V3 of this structure and input the values of members V1 and V2. 
Finally, we add members of V1 and V2 and keep the results in V3 and print it.